requirements.txt
PyDrive2==1.17.0
python-dotenv==1.0.0
python-gnupg==0.5.1